[AD-543] change dataType parameter to int16_t

+++ src/odbc/include/ignite/odbc/meta/column_meta.h
@@ -86,7 +86,7 @@ class ColumnMeta {
    * @param dataType Data type.
    */
   ColumnMeta(const std::string& schemaName, const std::string& tableName,
-             const std::string& columnName, int8_t dataType)
+             const std::string& columnName, int16_t dataType)
       : schemaName(schemaName),
         tableName(tableName),
         columnName(columnName),
